requirements.txt
numpy==2.2.6
opencv-python==4.12.0.88
pandas==2.3.3
pillow==12.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tqdm==4.67.1
tzdata==2025.2
